requirements.txt
django==4.1
djangorestframework==3.13.1
celery==5.2.7
requests==2.28.1